requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
dnspython==2.7.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
pymongo==4.11.3
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
